Reorganize file structure to simplify combination of services.
